On second thought, the SoundSourceProxy should grab the security token.
